+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ UNRELEASED CHANGES
 * CIP-2364: Calculate and append VCC ripple correction gains to the HPS FSP configuration for the 16k fine channelizers
 * CIP-3092: Scaled Up VCC Devices Deployed By MCS to 8 and SLIM Devices Needed to Support The Additional VCC
 * CIP-3344: Added State Attribute Wait Loop to Monitor State Change to Desired State; Added State Monitor in Controller's Start and Stop Communicating Functions.
+* CIP-2814: Implement FspPstSubarray for scan configuration and refactoring shared CORR and PST FSP Subarray code
 * SKB-729: Fix Off command and TalonBoard healthState monitoring
   * SLIM Off command can run if SLIM was powered ON but failed configuration
   * Controller re-tries TalonBoard start_communicating when re-trying On command to restart InfluxDB and HPS Master polling loop